refactor(ui): move NotApplicable into Browse3
