+++ weave-js/src/components/UserLink.tsx
@@ -1,7 +1,7 @@
 import {ApolloClient, gql, useApolloClient} from '@apollo/client';
 import {Avatar, Popover, TooltipProps} from '@mui/material';
 import * as Colors from '@wandb/weave/common/css/color.styles';
-import {NotApplicable} from '@wandb/weave/components/PagePanelComponents/Home/Browse2/NotApplicable';
+import {NotApplicable} from '@wandb/weave/components/PagePanelComponents/Home/Browse3/NotApplicable';
 import React, {useEffect, useRef, useState} from 'react';
 import styled from 'styled-components';
